--- app/src/main/java/WorkoutRoutines/R.java
@@ -1,5 +1,0 @@
-package WorkoutRoutines;
-
-public class R {
-    public static Object layout;
-}
