+++ app/src/main/java/com/example/olympia/WorkoutRoutines/WorkoutRoutines/ChestWorkout.java
@@ -1,16 +1,15 @@
-package WorkoutRoutines;
+package com.example.olympia.WorkoutRoutines.WorkoutRoutines;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
-
 import com.example.olympia.R;
 
-public class LegWorkout extends AppCompatActivity {
+public class ChestWorkout extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_leg_workout);
+        setContentView(R.layout.activity_chest_workout);
     }
 }
