+++ app/src/main/java/com/example/olympia/WorkoutRoutines/WorkoutRoutines/LegWorkout.java
@@ -1,15 +1,16 @@
-package WorkoutRoutines;
+package com.example.olympia.WorkoutRoutines.WorkoutRoutines;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
 import com.example.olympia.R;
 
-public class ChestWorkout extends AppCompatActivity {
+public class LegWorkout extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_chest_workout);
+        setContentView(R.layout.activity_leg_workout);
     }
 }
