+++ app/src/main/java/com/example/olympia/WorkoutRoutines/WorkoutRoutines/R.java
@@ -1,0 +1,5 @@
+package com.example.olympia.WorkoutRoutines.WorkoutRoutines;
+
+public class R {
+    public static Object layout;
+}
