+++ app/src/main/res/layout/activity_chest_workout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="WorkoutRoutines.ChestWorkout"
+    tools:context="com.example.olympia.WorkoutRoutines.WorkoutRoutines.ChestWorkout"
     tools:ignore="SpeakableTextPresentCheck">
 
     <LinearLayout
